backend/requirements.txt
anyio==3.3.4
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
black==21.9b0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
flake8==4.0.1
h11==0.12.0
idna==3.3
mccabe==0.6.1
motor==2.5.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
pymongo==3.12.1
python-dotenv==0.19.1
python-multipart==0.0.4
regex==2021.10.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
